pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "depadvisor"
version = "0.1.0"
description = "AI-powered dependency update advisor built with LangGraph"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Saurabh" }]
keywords = ["dependencies", "security", "ai", "langgraph", "devops"]

dependencies = [
    "langgraph>=1.0.0",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "packaging>=24.0",
    "python-dotenv>=1.0",
    "langsmith>=0.7",
]

[project.optional-dependencies]
server = ["fastapi>=0.115", "uvicorn>=0.30"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.22",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
depadvisor = "depadvisor.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/depadvisor"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["UP042"]  # Allow str+Enum pattern for Python 3.10 compat

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
